fixed warning about passing a double to an integer funtion
From: Alexander Petry <alexander.petry@itwm.fraunhofer.de>

git-svn-id: https://svn.itwm.fhg.de/svn/SDPA@5992 c3adf61f-8968-0410-94d0-83abfdf586cf

+++ trunk/sdpa/trunk/plugin/master.cpp
@@ -61,7 +61,7 @@ namespace drts
     if (m_cur_poll_interval < m_max_poll_interval)
     {
       m_cur_poll_interval +=
-        boost::posix_time::milliseconds(std::exp((double)m_poll_backoff_counter/3.0));
+        boost::posix_time::milliseconds((int)(std::exp((double)m_poll_backoff_counter/3.0)));
       if (m_cur_poll_interval > m_max_poll_interval)
         m_cur_poll_interval = m_max_poll_interval;
       ++m_poll_backoff_counter;
